Disturbance Model Running
Moved the model initialization outside the loop and fine tuned the weights and scaling factors to make the system stable and the Hessian Matrix small enough to not throw errors. Note in the review(mpc_obj) analyzer, this system is still not perfect because the Hessian is still a couple orders of magnitude larger for the inertial position states. You can't scale these too much or the cost function will no longer choose controls that move the aircraft towards the goal.