Updated Wind Field Model
It now scales the wind based on altitude. 500 m is the worst by the way not that we will fly that low. I was unsure which is your main file to run SMC. I know it isn't the SMC function, but I couldn't find where it was called. Anyway, I added to the struct in SMC. That's where you can initialize the smooth noise matrix and it will only be indexed here.